server/src/main/java/no/nav/abac/core/annotation/AbacMethodInterceptor.java
package no.nav.abac.core.annotation;

import no.nav.abac.core.annotation.attribute.AbacAttributePopulator;
import no.nav.abac.core.annotation.context.AbacContext;
import org.aopalliance.intercept.MethodInterceptor;
import org.aopalliance.intercept.MethodInvocation;

public class AbacMethodInterceptor implements MethodInterceptor {
    private final AbacContext context;
    private final AbacAttributePopulator populator;

    public AbacMethodInterceptor(AbacContext context, AbacAttributePopulator populator) {
        this.context = context;
        this.populator = populator;
    }

    @Override
    public Object invoke(MethodInvocation mi) throws Throwable {
        Abac abac = mi.getMethod().getAnnotation(Abac.class);

        populator.populate(context.getRequest(), abac);

        try {
            return mi.proceed();
        } finally {
            context.cleanUp();
        }
    }
}
